adapters/vox-parakeet/pyproject.toml
[project]
name = "vox-parakeet"
version = "0.1.1"
description = "NVIDIA Parakeet STT adapters for Vox"
requires-python = ">=3.11"
dependencies = [
    "vox>=0.1.0",
    "onnx-asr[gpu,hub]>=0.10.0; platform_machine == 'x86_64'",
    "onnx-asr[hub]>=0.10.0; platform_machine != 'x86_64'",
    "numpy>=1.26.0,<2.4",
    "soundfile>=0.13.1",
    "torch>=2.1.0",
    "nemo-toolkit[asr]>=2.7.0",
]

[project.entry-points."vox.adapters"]
parakeet = "vox_parakeet.adapter:ParakeetAdapter"
parakeet-nemo = "vox_parakeet.nemo_adapter:ParakeetNemoAdapter"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vox_parakeet"]
